requirements.txt
aiohttp==2.3.4
aiomysql==0.0.9
argh==0.26.2
async-generator==1.8
async-timeout==2.0.0
attrs==17.3.0
bencoder.pyx==1.2.0
chardet==3.0.4
click==6.7
Jinja2==2.10
Maga==3.0.0
markdown2==2.3.5
MarkupSafe==1.0
multidict==3.3.2
pathtools==0.1.2
peewee==2.10.2
peewee-async==0.5.10
pluggy==0.6.0
poyo==0.4.1
py==1.5.2
PyJWT==1.5.3
PyMySQL==0.7.11
pytest==3.3.0
pytest-asyncio==0.8.0
PyYAML==3.12
six==1.11.0
watchdog==0.8.3
yarl==0.15.0
